requirements.txt
langchain-core
langchain-community
langgraph
optimum
langchain_openai
python-dotenv
langchain-experimental 
tavily-python 
matplotlib
langchain-tavily
uvicorn
fastapi
azure-cosmos
PyPDF2
python-docx
langchain-text-splitters